Get rid of index.js, move headers fixture
