--- packages/js-dapi-client/lib/BlockHeadersProvider/index.js
@@ -1,3 +1,0 @@
-const BlockHeadersProvider = require('./BlockHeadersProvider');
-
-module.exports = BlockHeadersProvider;
